+++ packages/deepsea-tools/src/utils/treeToFibers.ts
@@ -23,10 +23,10 @@ export type Fiber<T> = T & {
 /**
  * 将树转换为 fiber
  * @param tree 要转换的树
- * @returns 返回转换后的 fiber
+ * @returns 如果树为空，则返回 null，否则返回转换后的 fiber
  */
-export function treeToFiber<T>(tree: Node<T>[]): Fiber<T> {
-    if (tree.length === 0) throw new Error("树不能为空")
+export function treeToFiber<T>(tree: Node<T>[]): Fiber<T> | null {
+    if (tree.length === 0) return null
     let first: Fiber<T>
     function createFiber(tree: Node<T>[], parent: Fiber<T> | null): void {
         let prev: Fiber<T> | null = null